# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Store Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: barcode
          in: query
          required: false
          schema:
            type: string
        - name: lowStock
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
...
  /invoices:
    get:
      operationId: getInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: startDate
          in: query
          required: false
          schema:
            type: string
        - name: endDate
          in: query
          required: false
          schema:
            type: string
        - name: customerId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice deleted
...
  /licenses:
    get:
      operationId: getLicenses
      tags: [licenses]
      summary: List licenses
      responses:
        "200":
          description: List of licenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/License"
    post:
      operationId: createLicense
      tags: [licenses]
      summary: Create license
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseInput"
      responses:
        "201":
          description: License created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
  /licenses/{id}:
    patch:
      operationId: updateLicense
      tags: [licenses]
      summary: Update license
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseUpdate"
      responses:
        "200":
          description: License updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
    delete:
      operationId: deleteLicense
      tags: [licenses]
      summary: Delete license
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: License deleted
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, price, categoryId, stock, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        costPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        stock:
          type: integer
        minStock:
          type: integer
        barcode:
          type: ["string", "null"]
        unit:
          type: string
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, price, categoryId, stock]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        costPrice:
          type: number
        categoryId:
          type: integer
        stock:
          type: integer
        minStock:
          type: integer
        barcode:
          type: string
        unit:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        costPrice:
          type: number
        categoryId:
          type: integer
        stock:
          type: integer
        minStock:
          type: integer
        barcode:
          type: string
        unit:
          type: string
...
    InvoiceItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice, total]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        discount:
          type: number
        total:
          type: number
    InvoiceItemInput:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId:
          type: integer
        quantity:
          type: number
        unitPrice:
          type: number
        discount:
          type: number
    Invoice:
      type: object
      required: [id, invoiceNumber, total, status, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card, transfer, credit]
        status:
          type: string
          enum: [draft, paid, cancelled]
        notes:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
    InvoiceDetail:
      type: object
      required: [id, invoiceNumber, total, status, createdAt, items]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card, transfer, credit]
        status:
          type: string
          enum: [draft, paid, cancelled]
        notes:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
    InvoiceInput:
      type: object
      required: [items, paymentMethod]
      properties:
        customerId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
        discount:
          type: number
        tax:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card, transfer, credit]
        status:
          type: string
          enum: [draft, paid, cancelled]
        notes:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [draft, paid, cancelled]
        notes:
          type: string
        discount:
          type: number
...
    License:
      type: object
      required: [id, name, licenseKey, expiryDate, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        licenseKey:
          type: string
        vendor:
          type: ["string", "null"]
        expiryDate:
          type: string
        status:
          type: string
          enum: [active, expired, cancelled]
        notes:
          type: ["string", "null"]
        cost:
          type: ["number", "null"]
        createdAt:
          type: string
    LicenseInput:
      type: object
      required: [name, licenseKey, expiryDate]
      properties:
        name:
          type: string
        licenseKey:
          type: string
        vendor:
          type: string
        expiryDate:
          type: string
        status:
          type: string
          enum: [active, expired, cancelled]
        notes:
          type: string
        cost:
          type: number
    LicenseUpdate:
      type: object
      properties:
        name:
          type: string
        licenseKey:
          type: string
        vendor:
          type: string
        expiryDate:
          type: string
        status:
          type: string
          enum: [active, expired, cancelled]
        notes:
          type: string
        cost:
          type: number
...