# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Store Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  costPrice?: number;
  categoryId: number;
  stock: number;
  minStock?: number;
  barcode?: string;
  unit?: string;
}